Produce glTF objects

The glTF file is the JSON part of a glTF representation of a 3D scene. This function creates the R structure corresponding to one, and writes the binary buffer file for it.

as.gltf(x, ...)
# S3 method for mesh3d
as.gltf(x, result = list(), newScene = FALSE, dir = tempdir(), ...)

Arguments

x
An object to convert to a "gltf" object.

result
Optionally a previously produced "gltf" object; the new geometry will be added to it.

newScene
logical; if TRUE, add a new scene to result, otherwise try to add to the existing scene.

dir
Where to write the binary buffer file.

...
Other parameters.

Details

as.gltf is a generic function.

Value

A "gltf" object.

References

The specification of the glTF format: https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html

Author

Duncan Murdoch

Examples

cube <- rgl::cube3d(col = "red")
gltf <- as.gltf(cube)
rgl::as.mesh3d(gltf)