Print the sequence of matrices affecting an object.

This is used for debugging, to see how an object within nested rgl subscenes is affected by their user matrices.

matrixSequence(tag, scene = scene3d())
# S3 method for matrixSequence
print(x, n = 5, ...)

Arguments

tag

Which objects to report on?

scene

The rgl scene to work from.

x

Object to print.

n

In the print method, how many vertices and indices should be printed?

...

Ignored.

Value

A list containing records with entries as follows:

id

The object id

vertices

The object's vertices

indices

The object's indices

userMatrix

A list of user matrices affecting this object