Print gltf objects.

Prints "gltf" objects with various levels of detail.

# S3 method for gltf
print(x, verbose = FALSE, names = FALSE, ...)

Arguments

x

The object to print.

verbose

A character vector naming the sections that should be printed in full detail, or a logical value, with TRUE indicating all sections should be printed verbosely.

names

A character vector naming the sections for which object names should be printed, if present.

...

Currently ignored.

Details

The recognized section names for verbose are: "accessors", "asset", "scene", "scenes", "nodes", "buffers", "bufferViews", "meshes", "cameras", "materials", "textures" and "images". Unique prefixes will be recognized.

Value

The x object is printed and then returned invisibly.

Examples

samples <- "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/master/2.0"

gltf <- readGLB(paste0(samples, "/2CylinderEngine/glTF-Binary/2CylinderEngine.glb?raw=true"))

print(gltf, names = "meshes")
#> asset fields:
#>   GLtf version 2.0 file.
#>   Generated by COLLADA2GLTF.
#> Scenes ( 1 )
#> Nodes ( 82 )
#> Buffers ( 1 )
#> Bufferviews ( 2 )
#> Meshes ( 29 )
#>   0: Piston_123-844_0_Parts_1
#>   1: body_24
#>   2: body_23
#>   3: body_22
#>   4: body_21
#>   5: body_20
#>   6: Spring_Link__0_Parts_1
#>   7: body_19
#>   8: body_18
#>   9: body_17
#>   10: body_16
#>   11: body_15
#>   12: body_14
#>   13: body_13
#>   14: body_12
#>   15: body_11
#>   16: body_10
#>   17: body_9
#>   18: body_8
#>   19: body_7
#>   20: body_6
#>   21: body_5
#>   22: body_4
#>   23: body
#>   24: body_3
#>   25: body_2
#>   26: body_1
#>   27: rod_123-699_0_Parts_1
#>   28: Lifter_123-923_0_Parts_1
#> Cameras ( 1 )
#> Accessors ( 102 )
#> Materials ( 34 )