Produce glTF objects

The glTF file is the JSON part of a glTF representation of a 3D scene. This function creates the R structure corresponding to one, and writes the binary buffer file for it.

as.gltf(x, ...)

# S3 method for default
as.gltf(x, y = NULL, z = NULL, vertices,
                            material = NULL,
                            normals = NULL,
                            texcoords = NULL,
                            points = NULL, segments = NULL,
                            triangles = NULL,
                            quads = NULL,
                            ...,
                            previous = list(),
                            newScene = FALSE,
                            dir = tempdir())

# S3 method for mesh3d
as.gltf(x, ...)

Arguments

x
An object to convert to a "gltf" object.

y, z
In the default method, combined with x to make coordinates. Any reasonable way of defining the coordinates is acceptable. See the function xyz.coords for details.

vertices
A 3 or 4 row matrix of Euclidean or homogeneous coordinates; takes precedence over x, y, z.

material
material properties for rendering

normals
normals at each vertex as a 3 or 4 column matrix

texcoords
texture coordinates at each vertex as a 2 column matrix

points
vector of indices of vertices to draw as points

segments
2 x n matrix of indices of vertices to draw as segments

triangles
3 x n matrix of indices of vertices to draw as triangles

quads
4 x n matrix of indices of vertices to draw as quads

...
Other parameters passed to the default method.

previous
Optionally a previously produced "gltf" object; the new geometry will be added to it.

newScene
logical; if TRUE, add a new scene to previous, otherwise try to add to the existing scene.

dir
Where to write the binary buffer file.

Details

as.gltf is a generic function.

Note

The arguments to the default method are similar to the arguments to the mesh3d function, with a few additional arguments after the dots.

Value

A "gltf" object.

References

The specification of the glTF format: https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html

Author

Duncan Murdoch

Examples

cube <- rgl::cube3d(col = "red")
gltf <- as.gltf(cube)
rgl::plot3d(gltf)