Debugging tool: show tags for objects in rgl scene

This function uses text3d to display the tags at the centre of the objects they label.

showTags(tags = NULL, ids = NULL,
         subscenes = ids3d("subscene", subscene = 0)$id,
         depth_test = "always", ...)

Arguments

tags, ids

If non-NULL, display only these tags and ids.

subscenes

Which subscenes to examine.

depth_test

The "depth_test" material property to use on the displayed tags. The default will put the tag in front of everything, so it won't be obscured by other objects.

...

Other arguments to pass to text3d.

Details

If selected objects don't have tags, they will be labelled using their id value instead.

Value

The rgl ids of the text objects added (one text object per tagged object, with tags as names).

Examples

example("plot3d", package = "rgl")
#> 
#> plot3d>   open3d()
#> glX 
#>  11 
#> 
#> plot3d>   x <- sort(rnorm(1000))
#> 
#> plot3d>   y <- rnorm(1000)
#> 
#> plot3d>   z <- rnorm(1000) + atan2(x, y)
#> 
#> plot3d>   plot3d(x, y, z, col = rainbow(1000))
showTags()